Latest Commits

+++ ElibraryManagement/bookdetails.aspx.cs
@@ -8,15 +8,21 @@ using System.Web.UI;
 using System.Web.UI.WebControls;
 using System.Data;
 using System.IO;
+using System.Reflection;
 
 namespace ElibraryManagement
 {
     public partial class bookdetails : System.Web.UI.Page
     {
         string strCon = ConfigurationManager.ConnectionStrings["con"].ConnectionString;
+        static string global_filepath;
+        static int global_actual_stock, global_current_stock, global_issued_books;
         protected void Page_Load(object sender, EventArgs e)
         {
-            fillAuthorPublisherValues();
+            if (!IsPostBack)
+            {
+                fillAuthorPublisherValues();
+            }
             displayGrid.DataBind();
         }
 
@@ -39,12 +45,28 @@ namespace ElibraryManagement
 
         protected void btnUpdate_Click(object sender, EventArgs e)
         {
+            if (checkIfBookExists())
+            {
+                updateBookbyID();
+            }
+            else
+            {
+                Response.Write("<script>alert('The Book is does not exists');</script>");
 
+            }
         }
 
         protected void btnDelete_Click(object sender, EventArgs e)
         {
-            deleteBook();
+            if (checkIfBookExists())
+            {
+                deleteBook();
+            }
+            else
+            {
+                Response.Write("<script>alert('The Book is does not exists');</script>");
+
+            }           
         }
 
         void fillAuthorPublisherValues()
@@ -115,6 +137,82 @@ namespace ElibraryManagement
             }
         }
 
+        void updateBookbyID()
+        {
+            try
+            {
+                int actual_stock = Convert.ToInt32(txt_actualStock.Text);
+                int current_stock = Convert.ToInt32(txt_currentStock.Text);
+                if(global_actual_stock == actual_stock)
+                {
+
+                }
+                else
+                {
+                    if(actual_stock < global_issued_books)
+                    {
+                        Response.Write("<script>alert('Actuall stock value cannot be less than the issued books');</script>");
+                        return;
+                    }
+                    else
+                    {
+                        current_stock = actual_stock - global_issued_books;
+                        txt_currentStock.Text = "" + current_stock;
+                    }
+                }
+
+                string genres = "";
+                foreach (int i in lbox_genre.GetSelectedIndices())
+                {
+                    genres = genres + lbox_genre.Items[i] + ",";
+                }
+                //genres = Adventure,Action,
+                genres = genres.Remove(genres.Length - 1);
+
+                string filePath = "/book_inventory/bookdetails.png";
+                string fileName = Path.GetFileName(FileUpload1.PostedFile.FileName);
+                if(fileName == "" || fileName == null)
+                {
+                    filePath = global_filepath;
+                }
+                else
+                {
+                    FileUpload1.SaveAs(Server.MapPath("book_inventory/" + fileName));
+                    filePath = "book_inventory/" + fileName;
+                }
+
+                SqlConnection con = new SqlConnection(strCon);
+                if (con.State == ConnectionState.Closed)
+                {
+                    con.Open();
+
+                }
+                SqlCommand cmd = new SqlCommand("UPDATE book_master_tbl set genre=@genre,author_name=@author_name,publisher_name=@publisher_name,publish_date=@publish_date,language=@language,edition=@edition,book_cost=@book_cost,no_of_pages=@no_of_pages,book_description=@book_description,actual_shock=@actual_shock,book_img_link=@book_img_link where book_id=@book_id OR book_name=@book_name", con);
+                cmd.Parameters.AddWithValue("@book_id", txt_bookID.Text);
+                cmd.Parameters.AddWithValue("@book_name", txt_bookName.Text);
+                cmd.Parameters.AddWithValue("@genre", genres);
+                cmd.Parameters.AddWithValue("@author_name", ddlAuthorName.SelectedValue);
+                cmd.Parameters.AddWithValue("@publisher_name", ddlPublisherName.SelectedValue);
+                cmd.Parameters.AddWithValue("@publish_date", txt_publishDate.Text);
+                cmd.Parameters.AddWithValue("@language", ddlLanguage.SelectedValue);
+                cmd.Parameters.AddWithValue("@edition", txt_Edition.Text);
+                cmd.Parameters.AddWithValue("@book_cost", txt_bookCost.Text);
+                cmd.Parameters.AddWithValue("@no_of_pages", txt_pages.Text);
+                cmd.Parameters.AddWithValue("@book_description", txt_bookDesc.Text);
+                cmd.Parameters.AddWithValue("@actual_shock", actual_stock.ToString());
+                cmd.Parameters.AddWithValue("@current_shock", current_stock.ToString());
+                cmd.Parameters.AddWithValue("@book_img_link", filePath);
+                cmd.ExecuteNonQuery();
+                con.Close();
+                displayGrid.DataBind();
+                Response.Write("<script>alert('Book Updated Succesfully');</script>");
+            }
+            catch (Exception ex)
+            {
+
+            }
+        }
+
         void addNewBook()
         {
             try
@@ -143,10 +241,10 @@ namespace ElibraryManagement
                 cmd.Parameters.AddWithValue("@bookID", txt_bookID.Text);
                 cmd.Parameters.AddWithValue("@book_name", txt_bookName.Text);
                 cmd.Parameters.AddWithValue("@genre", genres);
-                cmd.Parameters.AddWithValue("@author_name", ddlAuthorName.SelectedItem.Value);
-                cmd.Parameters.AddWithValue("@publisher_name", ddlPublisherName.SelectedItem.Value);
+                cmd.Parameters.AddWithValue("@author_name", ddlAuthorName.SelectedValue);
+                cmd.Parameters.AddWithValue("@publisher_name", ddlPublisherName.SelectedValue);
                 cmd.Parameters.AddWithValue("@publish_date", txt_publishDate.Text);
-                cmd.Parameters.AddWithValue("@language", ddlLanguage.SelectedItem.Value);
+                cmd.Parameters.AddWithValue("@language", ddlLanguage.SelectedValue);
                 cmd.Parameters.AddWithValue("@edition", txt_Edition.Text);
                 cmd.Parameters.AddWithValue("@book_cost", txt_bookCost.Text);
                 cmd.Parameters.AddWithValue("@no_of_pages", txt_pages.Text);
@@ -216,23 +314,39 @@ namespace ElibraryManagement
                     {
                         txt_bookID.Text = dr.GetValue(0).ToString();
                         txt_bookName.Text = dr.GetValue(1).ToString();
-                        lbox_genre.Text = dr.GetValue(2).ToString();
-                        ddlAuthorName.Text = dr.GetValue(3).ToString();
-                        ddlPublisherName.Text = dr.GetValue(4).ToString();
+                        ddlAuthorName.SelectedValue = dr.GetValue(3).ToString();
+                        ddlPublisherName.SelectedValue = dr.GetValue(4).ToString();
                         txt_publishDate.Text = dr.GetValue(5).ToString();
-                        ddlLanguage.Text = dr.GetValue(6).ToString();
+                        ddlLanguage.SelectedValue = dr.GetValue(6).ToString();
                         txt_Edition.Text = dr.GetValue(7).ToString();
                         txt_bookCost.Text = dr.GetValue(8).ToString();
-                        txt_pages.Text = dr.GetValue(9).ToString();
+                        txt_pages.Text = dr.GetValue(9).ToString().Trim();
                         txt_bookDesc.Text = dr.GetValue(10).ToString();
                         txt_actualStock.Text = dr.GetValue(11).ToString();
-                        txt_currentStock.Text = dr.GetValue(11).ToString();
+                        txt_currentStock.Text = dr.GetValue(12).ToString(); 
+                        txt_issuedBooks.Text = ""+(Convert.ToInt32(dr.GetValue(11).ToString()) - Convert.ToInt32(dr.GetValue(12).ToString()));
+                        lbox_genre.ClearSelection();
+                        string[] genre = dr.GetValue(2).ToString().Trim().Split(',');
+                        for(int i=0; i < genre.Length; i++)
+                        {
+                            for(int j = 0; j < lbox_genre.Items.Count; j++)
+                            {
+                                if (lbox_genre.Items[j].ToString() == genre[i])
+                                {
+                                    lbox_genre.Items[j].Selected = true;
+                                }
+                            }
+                        }
+                        global_actual_stock = Convert.ToInt32(dr.GetValue(11).ToString().Trim());
+                        global_current_stock = Convert.ToInt32(dr.GetValue(12).ToString().Trim());
+                        global_issued_books = global_actual_stock - global_current_stock;
+                        global_filepath = dr.GetValue(13).ToString();
                     }
                 }
 
                 else
                 {
-                    Response.Write("<script>alert('Invalid credentials');</script>");
+                    Response.Write("<script>alert('Invalid Book ID');</script>");
                 }
 
             }
@@ -255,6 +369,7 @@ namespace ElibraryManagement
             txt_bookDesc.Text = "";
             txt_actualStock.Text = "";
             txt_currentStock.Text = "";
+            lbox_genre.ClearSelection();
         }
     }
 }
